lifi widget color
