+++ packages/ui/app/dashboard/page.tsx
@@ -16,8 +16,8 @@ import type { LoopRowData } from '@ui/components/dashboards/LoopTable';
 import LoopTable from '@ui/components/dashboards/LoopTable';
 import type { SupplyRowData } from '@ui/components/dashboards/SupplyTable';
 import SupplyTable from '@ui/components/dashboards/SupplyTable';
-import ManageDialog from '@ui/components/dialogs/manage';
-import type { ActiveTab } from '@ui/components/dialogs/manage';
+import ManageDialog from '@ui/components/dialogs/ManageMarket';
+import type { ActiveTab } from '@ui/components/dialogs/ManageMarket';
 import NetworkSelector from '@ui/components/markets/NetworkSelector';
 import { FLYWHEEL_TYPE_MAP, pools } from '@ui/constants/index';
 import { useMultiIonic } from '@ui/context/MultiIonicContext';
