+++ packages/ui/app/market/details/[asset]/page.tsx
@@ -52,7 +52,7 @@ import {
 // const COLORS = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042'];
 import { INFO } from '@ui/constants/index';
 
-import Swap from '@ui/components/dialogs/manage/Swap';
+import Swap from '@ui/components/dialogs/ManageMarket/Swap';
 import { handleSwitchOriginChain } from '@ui/utils/NetworkChecker';
 import { MarketData } from '@ui/types/TokensDataMap';
 import { useFusePoolData } from '@ui/hooks/useFusePoolData';
@@ -62,7 +62,7 @@ import { Address, formatEther, formatUnits } from 'viem';
 import { useBorrowCapsDataForAsset } from '@ui/hooks/fuse/useBorrowCapsDataForAsset';
 import { useSupplyCapsDataForAsset } from '@ui/hooks/fuse/useSupplyCapsDataForPool';
 import BorrowAmount from '@ui/components/markets/BorrowAmount';
-import ManageDialog from '@ui/components/dialogs/manage';
+import ManageDialog from '@ui/components/dialogs/ManageMarket';
 import { useAssetChartData } from '@ui/hooks/useAssetChartData';
 import ChartWithDateRange from '@ui/components/markets/ChartWithDateRange';
 import ResultHandler from '@ui/components/ResultHandler';
