+++ packages/ui/app/stake/page.tsx
@@ -23,7 +23,7 @@ import {
   useWalletClient
 } from 'wagmi';
 
-import SliderComponent from '@ui/components/dialogs/manage/Slider';
+import SliderComponent from '@ui/components/dialogs/ManageMarket/Slider';
 import MaxDeposit from '@ui/components/MaxDeposit';
 import ResultHandler from '@ui/components/ResultHandler';
 import ClaimRewards from '@ui/components/stake/ClaimRewards';
