+++ packages/ui/components/dialogs/ManageMarket/Slider.tsx
