+++ packages/ui/components/dialogs/ManageMarket/tabs/BorrowTab.tsx
@@ -1,36 +1,38 @@
 import { useEffect, useMemo } from 'react';
 
+import { Info } from 'lucide-react';
 import { formatUnits } from 'viem';
 
 import MaxDeposit from '@ui/components/MaxDeposit';
+import ResultHandler from '@ui/components/ResultHandler';
+import { Alert, AlertDescription } from '@ui/components/ui/alert';
 import { Button } from '@ui/components/ui/button';
+import MemoizedUtilizationStats from '@ui/components/UtilizationStats';
 import {
   HFPStatus,
   TransactionType,
   useManageDialogContext
 } from '@ui/context/ManageDialogContext';
+import { useBorrow } from '@ui/hooks/market/useBorrow';
 import { useHealth } from '@ui/hooks/market/useHealth';
-import { useRepay } from '@ui/hooks/market/useRepay';
-import { useMaxRepayAmount } from '@ui/hooks/useMaxRepayAmount';
+import { useMaxBorrowAmount } from '@ui/hooks/useMaxBorrowAmount';
 
-import StatusAlerts from './StatusAlerts';
-import TransactionStepsHandler from './TransactionStepsHandler';
-import ResultHandler from '../../ResultHandler';
-import MemoizedUtilizationStats from '../../UtilizationStats';
+import StatusAlerts from '../StatusAlerts';
+import TransactionStepsHandler from '../TransactionStepsHandler';
 
-interface RepayTabProps {
+interface BorrowTabProps {
   capAmount: number;
   totalAmount: number;
   capFiat: number;
   totalFiat: number;
 }
 
-const RepayTab = ({
+const BorrowTab = ({
   capAmount,
   totalAmount,
   capFiat,
   totalFiat
-}: RepayTabProps) => {
+}: BorrowTabProps) => {
   const {
     selectedMarketData,
     resetTransactionSteps,
@@ -43,32 +45,44 @@ const RepayTab = ({
     isSliding
   } = useManageDialogContext();
 
-  const { data: maxAmount, isLoading: isLoadingMax } = useMaxRepayAmount(
+  const { data: maxAmount, isLoading: isLoadingMax } = useMaxBorrowAmount(
     selectedMarketData,
+    comptrollerAddress,
     chainId
   );
 
   const {
     isWaitingForIndexing,
-    repayAmount,
+    borrowAmount,
     isPolling,
-    currentBorrowAmountAsFloat,
+    borrowLimits,
+    isUnderMinBorrow,
     amount,
     setAmount,
     amountAsBInt
-  } = useRepay({
+  } = useBorrow({
     selectedMarketData,
-    chainId
+    chainId,
+    comptrollerAddress
   });
 
-  const { healthFactor, hfpStatus } = useHealth({
-    comptrollerAddress,
-    cToken: selectedMarketData.cToken,
-    activeTab: 'repay',
-    amount: amountAsBInt,
-    exchangeRate: selectedMarketData.exchangeRate,
-    decimals: selectedMarketData.underlyingDecimals
-  });
+  const { isLoadingPredictedHealthFactor, healthFactor, hfpStatus } = useHealth(
+    {
+      comptrollerAddress,
+      cToken: selectedMarketData.cToken,
+      activeTab: 'borrow',
+      amount: amountAsBInt,
+      exchangeRate: selectedMarketData.exchangeRate,
+      decimals: selectedMarketData.underlyingDecimals
+    }
+  );
+
+  const isDisabled =
+    !amount ||
+    amountAsBInt === 0n ||
+    isLoadingPredictedHealthFactor ||
+    hfpStatus === HFPStatus.CRITICAL ||
+    hfpStatus === HFPStatus.UNKNOWN;
 
   useEffect(() => {
     setPredictionAmount(amountAsBInt);
@@ -76,14 +90,14 @@ const RepayTab = ({
   }, [amountAsBInt]);
 
   const transactionSteps = useMemo(() => {
-    return getStepsForTypes(TransactionType.REPAY);
+    return getStepsForTypes(TransactionType.BORROW);
   }, [getStepsForTypes]);
 
   return (
     <div className="space-y-4 pt-4">
       <MaxDeposit
         max={formatUnits(
-          maxAmount ?? 0n,
+          maxAmount?.bigNumber ?? 0n,
           selectedMarketData.underlyingDecimals
         )}
         isLoading={isLoadingMax || isPolling}
@@ -91,34 +105,59 @@ const RepayTab = ({
         tokenName={selectedMarketData.underlyingSymbol}
         handleInput={(val?: string) => setAmount(val ?? '')}
         chain={chainId}
-        headerText="Repay Amount"
+        headerText="Borrow Amount"
         decimals={selectedMarketData.underlyingDecimals}
         showUtilizationSlider
-        hintText="Max Repay"
+        hintText="Max Borrow"
       />
+
+      {isUnderMinBorrow && (
+        <Alert
+          variant="default"
+          className="py-2 border-0 bg-opacity-90"
+        >
+          <div className="flex items-center">
+            <Info className="mr-2 h-4 w-4 text-white" />
+            <AlertDescription>
+              Amount must be greater than minimum borrow amount (
+              {borrowLimits.min} {selectedMarketData.underlyingSymbol})
+            </AlertDescription>
+          </div>
+        </Alert>
+      )}
 
       <StatusAlerts
         status={hfpStatus}
-        availableStates={[HFPStatus.CRITICAL]}
+        availableStates={[
+          HFPStatus.CRITICAL,
+          HFPStatus.UNKNOWN,
+          HFPStatus.WARNING
+        ]}
       />
 
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-4 lg:gap-x-8">
         <div className="space-y-4 content-center">
           <div className="flex justify-between text-xs text-gray-400">
+            <span>MIN BORROW</span>
+            <span>{parseFloat(borrowLimits.min).toFixed(6)}</span>
+          </div>
+
+          <div className="flex justify-between text-xs text-gray-400">
+            <span>MAX BORROW</span>
+            <span>{parseFloat(borrowLimits.max).toFixed(6)}</span>
+          </div>
+
+          <div className="flex justify-between text-xs text-gray-400">
             <span>CURRENTLY BORROWING</span>
             <div className="flex items-center">
-              <span className="text-error">
-                {updatedValues.borrowBalanceFrom}
-              </span>
+              <span>{updatedValues.borrowBalanceFrom}</span>
               <span className="mx-1">→</span>
               <ResultHandler
-                width={16}
+                isLoading={isSliding || isLoadingUpdatedAssets || isPolling}
                 height={16}
-                isLoading={isSliding || isLoadingUpdatedAssets}
+                width={16}
               >
-                <span className="text-accent">
-                  {updatedValues.borrowBalanceTo}
-                </span>
+                {updatedValues.borrowBalanceTo}
               </ResultHandler>
             </div>
           </div>
@@ -129,9 +168,9 @@ const RepayTab = ({
               <span>{updatedValues.borrowAPR?.toFixed(2)}%</span>
               <span className="mx-1">→</span>
               <ResultHandler
-                isLoading={isSliding || isLoadingUpdatedAssets || isPolling}
-                width={16}
                 height={16}
+                width={16}
+                isLoading={isSliding || isLoadingUpdatedAssets}
               >
                 {updatedValues.updatedBorrowAPR?.toFixed(2)}%
               </ResultHandler>
@@ -175,18 +214,16 @@ const RepayTab = ({
       ) : (
         <Button
           className="w-full bg-accent hover:bg-accent/80"
-          disabled={
-            !amount || amountAsBInt === 0n || !currentBorrowAmountAsFloat
-          }
-          onClick={repayAmount}
+          disabled={isDisabled || !!isUnderMinBorrow || isWaitingForIndexing}
+          onClick={borrowAmount}
         >
           {isWaitingForIndexing
             ? 'Updating Balances...'
-            : `Repay ${selectedMarketData.underlyingSymbol}`}
+            : `Borrow ${selectedMarketData.underlyingSymbol}`}
         </Button>
       )}
     </div>
   );
 };
 
-export default RepayTab;
+export default BorrowTab;
