+++ packages/ui/components/dialogs/ManageMarket/tabs/RepayTab.tsx
@@ -1,38 +1,36 @@
 import { useEffect, useMemo } from 'react';
 
-import { Info } from 'lucide-react';
 import { formatUnits } from 'viem';
 
 import MaxDeposit from '@ui/components/MaxDeposit';
-import { Alert, AlertDescription } from '@ui/components/ui/alert';
+import ResultHandler from '@ui/components/ResultHandler';
 import { Button } from '@ui/components/ui/button';
+import MemoizedUtilizationStats from '@ui/components/UtilizationStats';
 import {
   HFPStatus,
   TransactionType,
   useManageDialogContext
 } from '@ui/context/ManageDialogContext';
-import { useBorrow } from '@ui/hooks/market/useBorrow';
 import { useHealth } from '@ui/hooks/market/useHealth';
-import { useMaxBorrowAmount } from '@ui/hooks/useMaxBorrowAmount';
+import { useRepay } from '@ui/hooks/market/useRepay';
+import { useMaxRepayAmount } from '@ui/hooks/useMaxRepayAmount';
 
-import StatusAlerts from './StatusAlerts';
-import TransactionStepsHandler from './TransactionStepsHandler';
-import ResultHandler from '../../ResultHandler';
-import MemoizedUtilizationStats from '../../UtilizationStats';
+import StatusAlerts from '../StatusAlerts';
+import TransactionStepsHandler from '../TransactionStepsHandler';
 
-interface BorrowTabProps {
+interface RepayTabProps {
   capAmount: number;
   totalAmount: number;
   capFiat: number;
   totalFiat: number;
 }
 
-const BorrowTab = ({
+const RepayTab = ({
   capAmount,
   totalAmount,
   capFiat,
   totalFiat
-}: BorrowTabProps) => {
+}: RepayTabProps) => {
   const {
     selectedMarketData,
     resetTransactionSteps,
@@ -45,44 +43,32 @@ const BorrowTab = ({
     isSliding
   } = useManageDialogContext();
 
-  const { data: maxAmount, isLoading: isLoadingMax } = useMaxBorrowAmount(
+  const { data: maxAmount, isLoading: isLoadingMax } = useMaxRepayAmount(
     selectedMarketData,
-    comptrollerAddress,
     chainId
   );
 
   const {
     isWaitingForIndexing,
-    borrowAmount,
+    repayAmount,
     isPolling,
-    borrowLimits,
-    isUnderMinBorrow,
+    currentBorrowAmountAsFloat,
     amount,
     setAmount,
     amountAsBInt
-  } = useBorrow({
+  } = useRepay({
     selectedMarketData,
-    chainId,
-    comptrollerAddress
+    chainId
   });
 
-  const { isLoadingPredictedHealthFactor, healthFactor, hfpStatus } = useHealth(
-    {
-      comptrollerAddress,
-      cToken: selectedMarketData.cToken,
-      activeTab: 'borrow',
-      amount: amountAsBInt,
-      exchangeRate: selectedMarketData.exchangeRate,
-      decimals: selectedMarketData.underlyingDecimals
-    }
-  );
-
-  const isDisabled =
-    !amount ||
-    amountAsBInt === 0n ||
-    isLoadingPredictedHealthFactor ||
-    hfpStatus === HFPStatus.CRITICAL ||
-    hfpStatus === HFPStatus.UNKNOWN;
+  const { healthFactor, hfpStatus } = useHealth({
+    comptrollerAddress,
+    cToken: selectedMarketData.cToken,
+    activeTab: 'repay',
+    amount: amountAsBInt,
+    exchangeRate: selectedMarketData.exchangeRate,
+    decimals: selectedMarketData.underlyingDecimals
+  });
 
   useEffect(() => {
     setPredictionAmount(amountAsBInt);
@@ -90,14 +76,14 @@ const BorrowTab = ({
   }, [amountAsBInt]);
 
   const transactionSteps = useMemo(() => {
-    return getStepsForTypes(TransactionType.BORROW);
+    return getStepsForTypes(TransactionType.REPAY);
   }, [getStepsForTypes]);
 
   return (
     <div className="space-y-4 pt-4">
       <MaxDeposit
         max={formatUnits(
-          maxAmount?.bigNumber ?? 0n,
+          maxAmount ?? 0n,
           selectedMarketData.underlyingDecimals
         )}
         isLoading={isLoadingMax || isPolling}
@@ -105,59 +91,34 @@ const BorrowTab = ({
         tokenName={selectedMarketData.underlyingSymbol}
         handleInput={(val?: string) => setAmount(val ?? '')}
         chain={chainId}
-        headerText="Borrow Amount"
+        headerText="Repay Amount"
         decimals={selectedMarketData.underlyingDecimals}
         showUtilizationSlider
-        hintText="Max Borrow"
+        hintText="Max Repay"
       />
-
-      {isUnderMinBorrow && (
-        <Alert
-          variant="default"
-          className="py-2 border-0 bg-opacity-90"
-        >
-          <div className="flex items-center">
-            <Info className="mr-2 h-4 w-4 text-white" />
-            <AlertDescription>
-              Amount must be greater than minimum borrow amount (
-              {borrowLimits.min} {selectedMarketData.underlyingSymbol})
-            </AlertDescription>
-          </div>
-        </Alert>
-      )}
 
       <StatusAlerts
         status={hfpStatus}
-        availableStates={[
-          HFPStatus.CRITICAL,
-          HFPStatus.UNKNOWN,
-          HFPStatus.WARNING
-        ]}
+        availableStates={[HFPStatus.CRITICAL]}
       />
 
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-4 lg:gap-x-8">
         <div className="space-y-4 content-center">
           <div className="flex justify-between text-xs text-gray-400">
-            <span>MIN BORROW</span>
-            <span>{parseFloat(borrowLimits.min).toFixed(6)}</span>
-          </div>
-
-          <div className="flex justify-between text-xs text-gray-400">
-            <span>MAX BORROW</span>
-            <span>{parseFloat(borrowLimits.max).toFixed(6)}</span>
-          </div>
-
-          <div className="flex justify-between text-xs text-gray-400">
             <span>CURRENTLY BORROWING</span>
             <div className="flex items-center">
-              <span>{updatedValues.borrowBalanceFrom}</span>
+              <span className="text-error">
+                {updatedValues.borrowBalanceFrom}
+              </span>
               <span className="mx-1">→</span>
               <ResultHandler
-                isLoading={isSliding || isLoadingUpdatedAssets || isPolling}
-                height={16}
                 width={16}
+                height={16}
+                isLoading={isSliding || isLoadingUpdatedAssets}
               >
-                {updatedValues.borrowBalanceTo}
+                <span className="text-accent">
+                  {updatedValues.borrowBalanceTo}
+                </span>
               </ResultHandler>
             </div>
           </div>
@@ -168,9 +129,9 @@ const BorrowTab = ({
               <span>{updatedValues.borrowAPR?.toFixed(2)}%</span>
               <span className="mx-1">→</span>
               <ResultHandler
-                height={16}
+                isLoading={isSliding || isLoadingUpdatedAssets || isPolling}
                 width={16}
-                isLoading={isSliding || isLoadingUpdatedAssets}
+                height={16}
               >
                 {updatedValues.updatedBorrowAPR?.toFixed(2)}%
               </ResultHandler>
@@ -214,16 +175,18 @@ const BorrowTab = ({
       ) : (
         <Button
           className="w-full bg-accent hover:bg-accent/80"
-          disabled={isDisabled || !!isUnderMinBorrow || isWaitingForIndexing}
-          onClick={borrowAmount}
+          disabled={
+            !amount || amountAsBInt === 0n || !currentBorrowAmountAsFloat
+          }
+          onClick={repayAmount}
         >
           {isWaitingForIndexing
             ? 'Updating Balances...'
-            : `Borrow ${selectedMarketData.underlyingSymbol}`}
+            : `Repay ${selectedMarketData.underlyingSymbol}`}
         </Button>
       )}
     </div>
   );
 };
 
-export default BorrowTab;
+export default RepayTab;
