+++ packages/ui/context/ManageDialogContext.tsx
@@ -14,7 +14,7 @@ import debounce from 'lodash/debounce';
 import { type Address, formatUnits } from 'viem';
 import { useChainId } from 'wagmi';
 
-import type { TransactionStep } from '@ui/components/dialogs/manage/TransactionStepsHandler';
+import type { TransactionStep } from '@ui/components/dialogs/ManageMarket/TransactionStepsHandler';
 import { useMultiIonic } from '@ui/context/MultiIonicContext';
 import useUpdatedUserAssets from '@ui/hooks/ionic/useUpdatedUserAssets';
 import type { MarketData } from '@ui/types/TokensDataMap';
