+++ packages/ui/hooks/market/useRepay.ts
@@ -3,7 +3,7 @@ import { useCallback, useState, useMemo } from 'react';
 import { toast } from 'react-hot-toast';
 import { type Address, parseUnits, getContract, maxUint256 } from 'viem';
 
-import { useTransactionSteps } from '@ui/components/dialogs/manage/TransactionStepsHandler';
+import { useTransactionSteps } from '@ui/components/dialogs/ManageMarket/TransactionStepsHandler';
 import { INFO_MESSAGES } from '@ui/constants';
 import {
   TransactionType,
